First version with neural networks
The rabbits now have neural network brains that take in raycasts to food as eyesight input and use output layer to control movement. The raycasts are currently bugged, so the rabbits can only see directly in front of them. Interestingly, this means they learn faster because the input is simplified.